requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.4
django-countries==7.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.26.2
whitenoise==5.2.0
